requirements.txt
# TexCore — Backend Dependencies
# Versiones fijadas para builds reproducibles (actualizar intencionalmente con pip-compile)
# Última revisión: 2026-03-27

# Core Django
Django==5.2.7
asgiref==3.10.0
sqlparse==0.5.3

# Django REST Framework
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1

# CORS
django-cors-headers==4.9.0

# Base de datos SQL Server
mssql-django==1.6
pyodbc==5.2.0

# Servidor WSGI producción
gunicorn==25.1.0

# HTTP client (proxy a microservicios)
httpx==0.28.1
requests==2.32.5

# JWT Blacklist — requiere: python manage.py migrate
# Habilitado en settings.py: rest_framework_simplejwt.token_blacklist

# OpenAPI / Documentación automática
drf-spectacular==0.28.0

# Testing
factory-boy==3.3.3
coverage==7.8.0
